pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "distrostrap"
version = "0.1.0"
description = "Install any Linux distro onto a target partition from a running host system"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
distrostrap = "distrostrap.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
